merged_backend/requirements.txt
# FastAPI и зависимости
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
psycopg2-binary>=2.9.10

# Pydantic
pydantic==2.9.1
pydantic-settings==2.4.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP клиент для внешних API
httpx==0.27.2

# Environment variables
python-dotenv==1.0.0
